Move the initialization of the global notification service
We were initializing this in package load, because at one point the
APIs to do so had to be on the UI thread. At this point the only
remaining use of this service is for the legacy solution crawler,
so there's no point in initializing it until the service is actually
constructed. And since everything is now free threaded, it's easy to
move.
